wip: refactoring Login Form
