--- src/app/activate-page/activate.types.ts
@@ -1,4 +1,0 @@
-export interface IAuth {
-  email: string;
-  password: string;
-}
