+++ src/app/login/login.module.scss
@@ -4,52 +4,3 @@
 	font-weight: 600;
 	color: var(--primary-color-dark);
 }
-.alreadyHaveAccount {
-	text-align: center;
-	margin-bottom: 24px;
-}
-.innerForm {
-	max-width: 450px;
-	min-height: 450px;
-	margin: 0 auto;
-	margin-top: 80px;
-	padding: 36px;
-	border: 1px solid var(--primary-color);
-	border-radius: 12px;
-}
-.innerForm_paragraph {
-	margin-bottom: 24px;
-	font-size: 22px;
-	font-weight: 700;
-}
-.login {
-	color: var(--primary-color-light);
-	cursor: pointer;
-}
-.innerLine {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-
-	.line {
-		height: 1px;
-		width: 152px;
-	}
-}
-.innerImg {
-	display: flex;
-	justify-content: center;
-	margin-top: 24px;
-	img {
-		cursor: pointer;
-	}
-}
-.faggotPassword {
-	display: flex;
-	justify-content: space-between;
-}
-.faggotPassword_inner {
-	color: var(--primary-color-light);
-	cursor: pointer;
-}
-
